+++ .github/workflows/ci.yml
@@ -15,12 +15,8 @@ jobs:
 
       - name: Build
         run: |
-          export GOPATH=$(pwd)
-          cd src/monkey
           go build -v ./...
 
       - name: Test
         run: |
-          export GOPATH=$(pwd)
-          cd src/monkey
           go test -v ./...
