+++ .gitignore
@@ -1,2 +1,1 @@
-pkg
-src/monkey/monkey
+/cmd/monkey/monkey
