+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/freddiehaddad/monkey.interpreter
+
+go 1.20
