--- src/monkey/go.mod
@@ -1,3 +1,0 @@
-module monkey
-
-go 1.20
